First refactoring to discharge a bit the Resolver.
